BIT/requirements.txt
snowflake-connector-python==3.13.2
python-dotenv==1.0.1